Permissions and role

+++ src/main/java/com/yawarSoft/Modules/Admin/Repositories/PermissionRepository.java
@@ -3,5 +3,5 @@ package com.yawarSoft.Modules.Admin.Repositories;
 import com.yawarSoft.Core.Entities.PermissionEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PermissionRepository extends JpaRepository<PermissionEntity, Long> {
+public interface PermissionRepository extends JpaRepository<PermissionEntity, Integer> {
 }
